[ADD] formulario adaptado a la DB
